requirements.txt
python-dotenv
aiogram
openai
aiohttp
pillow
customtkinter
psutil